Move pinger pad tests
